Add files via upload
Environmental Informatics
Assignment 3

By Said El-Hawwat









	For this assignment we will be examining the displacement and temperature data from the acquisition of the specimen at The BEAST. Our numerical model is a linear thermal expansion problem. The sensors embedded into the concrete deck are slowly responsive to changes in ambient temperature. This is to be expected, as there are two inches of concrete in between the air and the sensors. Conversely, the sensors attached to the steel girders are much more responsive to changes in ambient temperature. Furthermore, steel is a much more deterministic material, so it’s easier to get more accurate results. Thus, our numerical model will attempt to describe the change in length of the steel girders as a function of temperature. 

We begin with the classic linear thermal expansion formula:
         ∆L=α ∆T L                                                              (1)
In our code, the variables will be discretized. However, for the sake of understanding, we will temporarily convert all the discrete differences in the formula to infinitesimally small differential elements:
dL=α L dT 
In order to express length as a function of temperature, let’s perform a separation of variables:
dL/L=α dT 
Upon integrating, we arrive at:
∫▒dL/L=∫▒〖α dT〗
ln⁡(L)=α T+C
, where C is an unknown constant. Let’s raise the natural number, e, to both sides of this equation:
L= e^(α T+C)
Applying the power rule:
                                                                              L= 〖A e〗^(α T)                                                                (2)

In the above formula, the length, L is a function of the temperature, T. The thermal coefficient, α, is equal to 7.2 〖E(-6)〗^o F^(-1) for steel We will solve it using initial conditions, L0 and T0. For our initial conditions, we will look at the first row in our length and temperature dataframes.
                                                                 1955.53= 〖A e〗^((7.6E-6)(40.24)) 
A=7953.1
Thus, our numerical model is:
                                                                          L= 〖7953.1 e〗^(α T)                                                        (3)